requirements.txt
# python==3.12.7
python-dotenv==1.1.0

fastapi==0.115.12
uvicorn==0.34.2

pydantic==2.11.3
pydantic-settings==2.6.1

langchain==0.3.24
langchain-openai==0.3.7
langchain-ollama==0.3.0
langchain-community==0.3.23
